Updating health stats

+++ Quit/ViewModel/QuitInfoVCViewModel.swift
@@ -11,7 +11,7 @@ import UserNotifications
 
 class QuitInfoVCViewModel {
     
-    func generateLocalNotif(title: String, body: String, minutes: Int, datePicker: Date) {
+    func generateLocalNotif(title: String, body: String, minutes: Double, datePicker: Date) {
         if #available(iOS 10.0, *) {
             let center = UNUserNotificationCenter.current()
             let content = UNMutableNotificationContent()
